remove unnecessary abstract class
